Reorganise and group the transfer-hook tests
Move transferHookExtraAccountMetas.test.ts into test/extensions/transferHook/ to match the existing extension-test layout, renaming it to extraAccountMetas.test.ts, and split the standalone default initialize/update instruction-builder tests into their own defaultExtraAccountMetaListInstructions.test.ts.

Group the tests by subject with describe blocks so each title reads as a natural sentence, and replace the bespoke createTestRpc helper with the shared createTestClient from _setup.ts. Also replace the new Array(n).fill(0) fixture helpers with a zero-filled Uint8Array, clearing the pre-existing no-new-array lint warnings.
